Rename files as object_registry
